pyproject.toml
[tool.poetry]
name = "trilhabackendjr-jun15"
version = "0.1.0"
description = ""
authors = ["FelipeSantiagoMenezes <felipemsanti05@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.32"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
taskipy = "^1.13.0"
ruff = "^0.5.7"
pytest-cov = "^5.0.0"
alembic = "^1.13.2"

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev src/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=src -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
